Move Float4 to headeronly
[ghstack-poisoned]
