+++ test/cpp/aoti_abi_check/test_dtype.cpp
@@ -2,6 +2,7 @@
 
 #include <c10/util/BFloat16-math.h>
 #include <c10/util/BFloat16.h>
+#include <torch/headeronly/util/Float4_e2m1fn_x2.h>
 #include <c10/util/Float8_e4m3fn.h>
 #include <c10/util/Float8_e4m3fnuz.h>
 #include <c10/util/Float8_e5m2.h>
@@ -80,6 +81,11 @@ TEST(TestDtype, TestFloat8_e5m2fnuz) {
   EXPECT_EQ(a - b, sub);
   EXPECT_EQ(a * b, mul);
   EXPECT_EQ(a / b, div);
+}
+
+TEST(TestDtype, TestFloat4) {
+  // not much you can do with this type, just make sure it compiles
+  torch::headeronly::Float4_e2m1fn_x2 a(5);
 }
 
 TEST(TestDtype, TestHalf) {
